+++ src/main/java/cc/cassian/raspberry/config/ModConfig.java
@@ -46,6 +46,7 @@ public class ModConfig {
     public boolean mapsWorkInInventory = false;
     public boolean create_emi = false;
     public boolean better_haste = true;
+    public boolean bannerlessRaiders = false;
 
     public static void load() {
         if (!Files.exists(configPath())) {