+++ src/main/java/cc/cassian/raspberry/mixin/minecraft/NoBannersMixin.java
@@ -1,0 +1,22 @@
+package cc.cassian.raspberry.mixin.minecraft;
+
+import cc.cassian.raspberry.ModCompat;
+import cc.cassian.raspberry.config.ModConfig;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.monster.PatrollingMonster;
+import net.minecraft.world.item.ItemStack;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+
+@Mixin(PatrollingMonster.class)
+public class NoBannersMixin {
+    @WrapOperation(method = "finalizeSpawn", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/monster/PatrollingMonster;setItemSlot(Lnet/minecraft/world/entity/EquipmentSlot;Lnet/minecraft/world/item/ItemStack;)V"))
+    private static void noBanners(PatrollingMonster instance, EquipmentSlot equipmentSlot, ItemStack stack, Operation<Void> original) {
+        if (!ModConfig.get().bannerlessRaiders) {
+            original.call(instance, equipmentSlot, stack);
+        }
+    }
+}